Add tab info custom color setting, related changes
- Encode custom color in existing tab info settings
- Some changes to ColorChooser/Setting to accomodate changing preview
  and no base color
- Fix tab setting not updating for default channel

+++ src/chatty/gui/DockedDialogHelper.java
@@ -195,7 +195,7 @@ public class DockedDialogHelper {
     
     public void loadTabSettings() {
         if (content instanceof DockStyledTabContainer) {
-            ((DockStyledTabContainer)content).setSettings(0, (int)settings.getLong("tabsMessage"), 0, 0, 0, -1);
+            ((DockStyledTabContainer)content).setSettings(0, settings.getLong("tabsMessage"), 0, 0, 0, -1);
         }
     }
